FileDownload component created allowing for download of files
